home

+++ src/main/java/com/ion/repository/FileRepo.java
@@ -3,5 +3,5 @@ package com.ion.repository;
 import com.ion.model.FileDTO;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FileRepo extends JpaRepository<FileDTO, Integer> {
+public interface FileRepo extends JpaRepository<FileDTO, Long> {
 }
